# [cascade] scheduling and memory improvements