- new benchmark job with matmul
- respects initial setting of CUDA_VISIBLE_DEVICES
- bugfix in worker claiming gpu even if not allowed to
- runner mem flushing: add new strategy of pre-flush
- runner mem flushing: configurable which strategy to choose
- rename scheduler/graph to scheduler/precompute
- fix impromer handling of task sequence assignments
- added preschedule option of fusing opportunities
- fine-grain the dataset publication command in case of non-trivial task sequences
- add preschedule impl of the fusion opportunities (heaviest-dataset crawls starting from sinks)
- add tests with graphs that would trigger some fusing

small benchmark -- matmul benchmark job with 8 nodes and 2**11 matrices:
 - before PR: ~2.5s on gpu (and ~6.5s on cpu)
 - executed without cascade as a for-cycle + final shmem write: ~2.0s (6.3s)
 - cumulative deserializations ~0.06s (savings of pre-flush strategy)
 - cumulative serializations ~0.4s (savings of fusing (which additionally saves ~.1s comms overhead)
=> improvement from 2.5s to 2.0s, ie, down to .8, with mod-noise-equivalent runtime to no-cascade


fusing and gpu interplay as by a change to assignment heuristic: now we consider "cpu task with a fusable gpu descendant" as a candidate _after_ all gpu-requiring tasks were considered, but _before_ cpu-only tasks are assigned, and we first attempt to assign those to gpu workers

I've executed a run of forecastbox's `bigtest` with this, and "still works" => good enough. Effect on runtime is indiscernible because its dominated by webmars noise. But regarding fusing statistics:
1/ the very first task gets fused, consisting of (get_initial conditions, empty_payload, run_as_earthkit, take), and publishes outputs (run_as_earthkit 0 to 6, take)
2/ then there is a bunch of unfused take & concat tasks
3/ and lastly there are two fused tasks, concat & quickplot, and convert2grib & output-grib
I'm a bit puzzled why the 3/ weren't also fused into something in 2/ and 1/, but this is still ~good. The main thing is probably fusing of initial_conditions and run_as_earthkit -- the remaining takes are quite a fan-out so no real gains obtainable by fusing